Refs: #446
Solves the missing Number of Pulses for fluoroscopic and rotational acquisitions on Philips Allura XPer systems.

Doesn't solve the missing end angles for rotational acquisitions in the RDSR
This seems to be a more difficult issue as there doesn't seem to be a one to one relation
between number of pulses and acquisition angle.