Refs #667
Changed more templates.

I wonder if it would be better to use the url reverse lookup option to prevent hard-coded urls in the templates and views.py.
Of course you then would have hard-coded names (but you can change the url, without having to change the templates and/or views.py).
Actually all the changes for this issue would become obsolete.
The only change needed would be in openremproject/urls.py.

A few items still need to be changed (mainly in the DICOM settings page).