Refs Issue #688
Makes it possible to automatically set display name (and modality) to the display name of an already known system if the device observer UID matches.

Default behaviour is not to apply this feature (backwards compatible)

Setting can be made in "home page Options", as the user will see the changes in the home page

If it is going to be implemented, the documentation has to be written.