Refs Issue #688
Fixed some Codacy complains.
Combined the messages for the admin settings into one message.

Btw: The home page settings are not available in the menu for a normal user.
So it can only be reached by typing the right URL. Should we place the menu option into the "user level config"?